Update mock test
